fix: arch review — remove LegacyJournal, rename NewJournal→Journal, centralise describe_* functions, stamp recorded_at
F10.3: Rename NewJournal to Journal everywhere, remove LegacyJournal
struct and dual-write path in apply_observations. The new Journal is
now the sole data model.

F10.2+F10.5: Create src/descriptions.rs collecting all describe_*
functions (describe_value, describe_density, describe_color,
describe_thermal_behavior, describe_thermal_observation) from
journal.rs, interaction.rs, and observation.rs. Removes duplicates.

F10.1: apply_observations now takes Res<Time> and stamps each
observation's recorded_at with elapsed milliseconds, so callers can
pass 0 and the ingestion system overwrites with real time.
